setting export and import has snackbar / refactor
